AddedMapingService

+++ src/CinemaPark.Business/Services/Implementations/GenreService.cs
@@ -1,8 +1,11 @@
-﻿using CinemaPark.Business.DTOs.GenreDtos;
+﻿using AutoMapper;
+using CinemaPark.Business.DTOs.GenreDtos;
+using CinemaPark.Business.DTOs.MovieDtos;
 using CinemaPark.Business.Services.Interfaces;
 using CinemaPark.Business.Utilities.Exceptions;
 using CinemaPark.Core.Entities;
 using CinemaPark.Core.Repositories;
+using Microsoft.EntityFrameworkCore;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -14,27 +17,29 @@ namespace CinemaPark.Business.Services.Implementations;
 public class GenreService : IGenreService
 {
     private readonly IGenreRepository _genreRepository;
-    public GenreService(IGenreRepository genreRepository)
+    private readonly IMapper _mapper;
+    public GenreService(IGenreRepository genreRepository,IMapper mapper)
     {
         _genreRepository = genreRepository;
+        _mapper = mapper;
     }
     public async Task CreateAsync(GenrePostDto genrePostDto)
     {
         if (genrePostDto is null) throw new NotFoundException(404,"Genre is not found");
-        Genre genre = new Genre()
-        {
-            Name= genrePostDto.Name,
-            IsDeleted=genrePostDto.IsDeleted,
-            CreatedDate=DateTime.UtcNow,
-            UpdatedDate=DateTime.UtcNow,
-        };
+        Genre genre=_mapper.Map<Genre>(genrePostDto);
+
+        genre.CreatedDate = DateTime.UtcNow;
+        genre.UpdatedDate = DateTime.UtcNow;
+
         await _genreRepository.InsertAsync(genre);
         await _genreRepository.CommitAsync();
     }
 
-    public async Task DeleteAsync(int id)
+    public async Task DeleteAsync(int id,GenreDeleteDto genreDeleteDto)
     {
-       var genre= await _genreRepository.GetByIdAsync(id);
+        if (id != genreDeleteDto.Id) throw new Exception("Id must be valid");
+        if (!await _genreRepository.IsExist(g => g.Id == genreDeleteDto.Id)) throw new NotFoundException(404, "Genre is not found");
+        var genre= await _genreRepository.GetByIdAsync(id);
         if(genre is null) throw new NotFoundException(404,"Genre is not found");
         _genreRepository.Delete(genre);
         await _genreRepository.CommitAsync();
@@ -47,12 +52,8 @@ public class GenreService : IGenreService
         var genres= await _genreRepository.GetAllAsync();
         foreach(var genre in genres)
         {
-            GenreGetDto dto = new GenreGetDto()
-            {
-                Id= genre.Id,
-                Name = genre.Name,
-            };
-            genreGetDtos.Add(dto);
+            GenreGetDto genreGetDto=_mapper.Map<GenreGetDto>(genre);
+            genreGetDtos.Add(genreGetDto);
         }
         return genreGetDtos;
     }
@@ -61,22 +62,19 @@ public class GenreService : IGenreService
     {
         var genre=await _genreRepository.GetByIdAsync(id);
         if(genre is null) throw new NotFoundException(404,"Genre is not found!");
-        GenreGetDto genreGetDto = new GenreGetDto()
-        {
-            Id = genre.Id,
-            Name = genre.Name,
-        };
+        GenreGetDto genreGetDto= _mapper.Map<GenreGetDto>(genre);
         return genreGetDto;
     }
 
-    public async Task UpdateAsync(int id, GenrePostDto genrePostDto)
+    public async Task UpdateAsync(int id, GenrePutDto genrePutDto)
     {
+        if (id != genrePutDto.Id) throw new Exception("id must be valid");
+        if (!await _genreRepository.IsExist(g => g.Id == genrePutDto.Id)) throw new NotFoundException(404, "Genre is not found");
         var currentGenre=await _genreRepository.GetByIdAsync(id);
         if (currentGenre is null) throw new NotFoundException(404, "Genre is not found!");
-        if (genrePostDto is null) throw new NotFoundException(404, "Genre is not found!");
-        currentGenre.UpdatedDate = DateTime.UtcNow;
-        currentGenre.Name= genrePostDto.Name;
-        currentGenre.IsDeleted=genrePostDto.IsDeleted;
+        if (genrePutDto is null) throw new NotFoundException(404, "Genre is not found!");
+        Genre genre=_mapper.Map(genrePutDto, currentGenre);
+        genre.UpdatedDate = DateTime.UtcNow;
         await _genreRepository.CommitAsync();
     }
